bsp starts

+++ ex00/inc/Fixed.hpp
@@ -6,7 +6,7 @@
 /*   By: jvalkama <jvalkama@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/20 15:09:43 by jvalkama          #+#    #+#             */
-/*   Updated: 2026/01/21 13:03:09 by jvalkama         ###   ########.fr       */
+/*   Updated: 2026/01/23 14:09:58 by jvalkama         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 class Fixed {
     private:
-        int                 _fp_val;
+        long long           _fp_val;
         static const int    _bits = 8;
 
     public:
